numeric extension move into core part 2
moved numeric/channel_numeric_operations.hpp into core
