+++ test/core/channel/Jamfile
@@ -1,6 +1,7 @@
 # Boost.GIL (Generic Image Library) - tests
 #
 # Copyright (c) 2018-2020 Mateusz Loskot <mateusz@loskot.net>
+# Copyright (c) 2021 Pranam Lashkari <plashkari628@gmail.com>
 #
 # Distributed under the Boost Software License, Version 1.0.
 # (See accompanying file LICENSE_1_0.txt or
@@ -15,6 +16,7 @@ run channel_traits.cpp ;
 run test_fixture.cpp ;
 run packed_channel_value.cpp ;
 run scoped_channel_value.cpp ;
+run channel_numeric_operations.cpp ;
 
 run algorithm_channel_arithmetic.cpp ;
 run algorithm_channel_convert.cpp ;
