+++ test/core/channel/channel_numeric_operations.cpp
@@ -1,5 +1,6 @@
 //
 // Copyright 2019-2020 Mateusz Loskot <mateusz at loskot dot net>
+// Copyright 2021 Pranam Lashkari <plashkari628@gmail.com>
 //
 // Distributed under the Boost Software License, Version 1.0
 // See accompanying file LICENSE_1_0.txt or copy at
