+++ test/extension/numeric/Jamfile
@@ -14,7 +14,6 @@ alias headers : [ generate_self_contained_headers extension/numeric ] ;
 
 compile-fail kernel_1d_fixed_even_size_fail.cpp ;
 
-run channel_numeric_operations.cpp ;
 run convolve.cpp ;
 run convolve_2d.cpp ;
 run convolve_cols.cpp ;
